% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRflow-package.R
\docType{package}
\name{covRflow-package}
\alias{covRflow}
\alias{covRflow-package}
\title{covRflow: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
