% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRflow.R
\name{covRflow}
\alias{covRflow}
\title{Display album covers in the Viewer pane}
\usage{
covRflow(data)
}
\arguments{
\item{data}{A tibble with the following necessary columns: \code{release_title} (title of the release), \code{artist} (artist) and \code{cover_image} (URL to an image of the release cover image). The tibble can be created with \code{\link[covRflow]{retrieve_covers_discogs}}. See examples for more details.}
}
\value{
The function does not have a return value.
}
\description{
Creates the interactive cover flow-esque display of album covers in the Viewer pane.
}
\examples{
\dontrun{
example_albums |>
  create_request_discogs(access_token = "thisisasecret") |>
  retrieve_covers_discogs() |>
  covRflow()
}
}
