% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_request.R
\name{create_request_discogs}
\alias{create_request_discogs}
\title{Create requests to retrieve covers from the Discogs API}
\usage{
create_request_discogs(data, access_token)
}
\arguments{
\item{data}{A tibble with the following columns: \itemize{
\item{\code{release_title}} {The title of the release, for example an album title.}
\item{\code{artist}} {The artist who created the release.}
\item{\code{year}} {The year of the release.}
\item{\code{format}} {The format of the release. Examples include "album", "LP", "CD", "Vinyl".}
} See}

\item{access_token}{A string containing your Discogs API token. Unfortunately, you have to identify yourself with your personal token to use the functionality necessery to search for album covers. If you do not have a token yet, go to \url{https://www.discogs.com/settings/developers} and sign in to your Discogs account (or create a new one for free) and click on "Generate new token".}
}
\value{
A tibble with a new list column called \code{request}.
}
\description{
Create requests to retrieve covers from the Discogs API
}
\examples{
\dontrun{
example_albums |>
  create_request_discogs(access_token = "thisisasecret") |>
  retrieve_covers_discogs() |>
  covRflow()
}
}
