% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_albums}
\alias{example_albums}
\title{Some randomly selected albums}
\format{
\subsection{\code{example_albums}}{

A tibble with 7 rows and 4 columns:
\describe{
\item{release_title}{The album title.}
\item{artist}{The performing artist.}
\item{year}{The year of release.}
\item{format}{The format. In this case all are albums but could also be "LP" or more specific, for example "CD" or "Vinyl".}
}
}
}
\source{
\url{https://100best.music.apple.com/us}
}
\usage{
example_albums
}
\description{
Albums were selected from Apple Music's top 100 albums of all time \url{https://100best.music.apple.com/us}.
}
\keyword{datasets}
