% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_covers.R
\name{retrieve_covers_discogs}
\alias{retrieve_covers_discogs}
\title{Retrieve cover URLs from the Discogs API}
\usage{
retrieve_covers_discogs(data)
}
\arguments{
\item{data}{A tibble containing a \code{request} column previously created with \link[covRflow]{create_request_discogs}.}
}
\value{
A tibble with new columns \code{response}, \code{body}, \code{id} and \code{cover_image}.
}
\description{
Retrieve album covers from the Discogs API with the requests created with \link[covRflow]{create_request_discogs}.
}
\examples{
\dontrun{
example_albums |>
  create_request_discogs(access_token = "thisisasecret") |>
  retrieve_covers_discogs() |>
  covRflow()
}
}
