% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{view_apache_license}
\alias{view_apache_license}
\title{Prints the full text of the Apache 2.0 license to the console}
\usage{
view_apache_license()
}
\value{
Prints the Apache 2.0 license to the console.
}
\description{
The HTML, CSS and JavaScript code necessary to produce the Coverflow in the Viewer panel is licensed under Apache 2.0.
Part of its requirements is to make the license available in packages using the licensed software.
This functions helps to fulfill that requirement.
}
\examples{
view_apache_license()
}
