% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRflow-package.R
\docType{package}
\name{covRflow-package}
\alias{covRflow}
\alias{covRflow-package}
\title{covRflow: Easily Retrieve Album Covers and Display them in the Style of the Famous Cover Flow}
\description{
Retrieve album covers from a web Application Programming Interface (API) and use them to create a cover flow in the viewer pane. Currently only the Discogs API \url{https://www.discogs.com/developers/#} is supported. Honestly just an excuse to play around with the Discogs API and the Viewer pane.
}
\author{
\strong{Maintainer}: Yannik Schermer \email{mail@yannik.science}

}
\keyword{internal}
